chore(cli): improve onlyPlugins file structure, include a unit test
