+++ packages/cli/src/lib/collect/collect-command.ts
@@ -5,7 +5,7 @@ import {
   collectAndPersistReports,
 } from '@code-pushup/core';
 import { CLI_NAME } from '../cli';
-import { onlyPluginsOption } from '../implementation/only-config-option';
+import { onlyPluginsOption } from '../implementation/only-plugins-options';
 
 export function yargsCollectCommandObject(): CommandModule {
   const command = 'collect';
