--- packages/cli/src/lib/implementation/only-config-option.ts
@@ -1,8 +1,0 @@
-import { Options } from 'yargs';
-
-export const onlyPluginsOption: Options = {
-  describe: 'List of plugins to run. If not set all plugins are run.',
-  type: 'array',
-  default: [],
-  coerce: (arg: string[]) => arg.flatMap(v => v.split(',')),
-};
