+++ packages/cli/src/lib/implementation/only-plugins-options.ts
@@ -1,3 +1,8 @@
-export interface OnlyPluginsOptions {
-  onlyPlugins: string[];
-}
+import { Options } from 'yargs';
+
+export const onlyPluginsOption: Options = {
+  describe: 'List of plugins to run. If not set all plugins are run.',
+  type: 'array',
+  default: [],
+  coerce: (arg: string[]) => arg.flatMap(v => v.split(',')),
+};
